# Conflicts:
#	app/Http/Controllers/VoteController.php
#	resources/assets/js/app/config/routers.js

+++ app/Http/Controllers/CommentController.php
@@ -9,7 +9,7 @@ use App\Http\Requests\CommentsRequest;
 use App\Models\Topic;
 use App\Http\Requests;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-
+use Gate;
 class CommentController extends ApiController
 {
     /**
@@ -18,10 +18,7 @@ class CommentController extends ApiController
      */
     protected function isCommentHasAnyChild(Comment $comment)
     {
-        if ($comment->comments()->get()) {
-            return true;
-        }
-        return false;
+        return $comment->comments()->exists();
     }
 
     /**
@@ -47,9 +44,7 @@ class CommentController extends ApiController
      */
     protected function isCommentBelongsToTopic(Topic $topic, Comment $comment)
     {
-        $topicWhichHasThisComment = $comment->topics()->get()->first();
-
-        return ($topicWhichHasThisComment && $topicWhichHasThisComment->id === $topic->id);
+        return !!$topic->comments()->find($comment->id);
     }
 
     /**
@@ -96,6 +91,8 @@ class CommentController extends ApiController
      */
     public function updateTopicComment(Topic $topic, Comment $comment, CommentsRequest $request)
     {
+        $this->authorize('updateTopicsComment', [$comment, $topic]);
+
         if ($this->isCommentBelongsToTopic($topic, $comment)) {
             $comment->update($request->all());
             return $this->setStatusCode(200)->respond($comment);
@@ -112,6 +109,8 @@ class CommentController extends ApiController
      */
     public function destroyTopicComment(Topic $topic, Comment $comment)
     {
+        $this->authorize('deleteTopicsComment', [$comment, $topic]);
+
         if ($this->isCommentBelongsToTopic($topic, $comment)) {
             $comment->delete();
             return $this->setStatusCode(204)->respond();
@@ -147,8 +146,9 @@ class CommentController extends ApiController
     {
         if ($this->isCommentBelongsToTopic($topic, $comment)) {
             $childComment = Comment::create($childCommentInput->all());
+            $topic->comments()->save($childComment);
             $childComment = $comment->comments()->save($childComment);
-            return $this->setStatusCode(200)->respond($childComment);
+            return $this->setStatusCode(201)->respond($childComment);
         } else {
             throw (new ModelNotFoundException)->setModel(Comment::class);
         }
@@ -184,6 +184,9 @@ class CommentController extends ApiController
         Comment $commentChild,
         CommentsRequest $request
     ) {
+
+        $this->authorize('updateTopicsComment', [$comment, $topic]);
+
         if ($this->isCommentBelongsToTopic($topic, $comment)
             && $this->isCommentChildBelongsToComment($comment, $commentChild)
         ) {
@@ -203,6 +206,8 @@ class CommentController extends ApiController
      */
     public function destroyTopicCommentChild(Topic $topic, Comment $comment, Comment $commentChild)
     {
+        $this->authorize('deleteTopicsComment', [$comment, $topic]);
+
         if ($this->isCommentBelongsToTopic($topic, $comment)
             && $this->isCommentChildBelongsToComment($comment, $commentChild)
         ) {
@@ -222,7 +227,7 @@ class CommentController extends ApiController
      */
     protected function isCommentBelongsToVote(Vote $vote, Comment $comment)
     {
-        $voteWhichHasThisComment = $comment->votes()->get()->first();
+        $voteWhichHasThisComment = $comment->commentable()->get()->first();
 
         return ($voteWhichHasThisComment && $voteWhichHasThisComment->id === $vote->id);
     }
@@ -288,6 +293,8 @@ class CommentController extends ApiController
      */
     public function updateVoteComment(Vote $vote, Comment $comment, CommentsRequest $request)
     {
+        $this->authorize('updateVotesComment', [$comment, $vote]);
+
         if ($this->isCommentBelongsToVote($vote, $comment)) {
             $comment->update($request->all());
             return $this->setStatusCode(200)->respond($comment);
@@ -304,6 +311,8 @@ class CommentController extends ApiController
      */
     public function destroyVoteComment(Vote $vote, Comment $comment)
     {
+        $this->authorize('deleteVotesComment', [$comment, $vote]);
+
         if ($this->isCommentBelongsToVote($vote, $comment)) {
             $comment->delete();
             return $this->setStatusCode(204)->respond();
@@ -340,7 +349,7 @@ class CommentController extends ApiController
         if ($this->isCommentBelongsToVote($vote, $comment)) {
             $childComment = Comment::create($childCommentInput->all());
             $childComment = $comment->comments()->save($childComment);
-            return $this->setStatusCode(200)->respond($childComment);
+            return $this->setStatusCode(201)->respond($childComment);
         } else {
             throw (new ModelNotFoundException)->setModel(Comment::class);
         }
@@ -376,6 +385,8 @@ class CommentController extends ApiController
         Comment $commentChild,
         CommentsRequest $request
     ) {
+        $this->authorize('updateVotesComment', [$comment, $vote]);
+
         if ($this->isCommentBelongsToVote($vote, $comment)
             && $this->isCommentChildBelongsToComment($comment, $commentChild)
         ) {
@@ -395,6 +406,8 @@ class CommentController extends ApiController
      */
     public function destroyVoteCommentChild(Vote $vote, Comment $comment, Comment $commentChild)
     {
+        $this->authorize('deleteVotesComment', [$comment, $vote]);
+
         if ($this->isCommentBelongsToVote($vote, $comment)
             && $this->isCommentChildBelongsToComment($comment, $commentChild)
         ) {
@@ -414,7 +427,7 @@ class CommentController extends ApiController
      */
     protected function isCommentBelongsToVoteItem(VoteItem $voteItem, Comment $comment)
     {
-        $voteItemWhichHasThisComment = $comment->voteItems()->get()->first();
+        $voteItemWhichHasThisComment = $comment->commentable()->get()->first();
 
         return ($voteItemWhichHasThisComment && $voteItemWhichHasThisComment->id === $voteItem->id);
     }
@@ -463,6 +476,8 @@ class CommentController extends ApiController
      */
     public function updateVoteItemComment(VoteItem $voteItem, Comment $comment, CommentsRequest $request)
     {
+        $this->authorize('updateVoteItemsComment', [$comment, $voteItem]);
+
         if ($this->isCommentBelongsToVoteItem($voteItem, $comment)) {
             $comment->update($request->all());
             return $this->setStatusCode(200)->respond($comment);
@@ -479,6 +494,8 @@ class CommentController extends ApiController
      */
     public function destroyVoteItemComment(VoteItem $voteItem, Comment $comment)
     {
+        $this->authorize('deleteVoteItemsComment', [$comment, $voteItem]);
+
         if ($this->isCommentBelongsToVoteItem($voteItem, $comment)) {
             $comment->delete();
             return $this->setStatusCode(204)->respond();
